Started work on Spell Selector
Changed getOrPopulate tag, as it didn't merge correctly.

Created a screens package for screens.
Moved screens and their handlers into it.

Created KeybindUtil to keep track of keybinds.

Changed Salem.java to reveive GUI packet (Needs to be moved externally to some kind of PacketUtil/Handler).

Started work on SpellPicker GUI.
Rendered item holder on screen, as well as the item the player has activated it with. (Locked to items that are an implement SpellCaster).

Fixed/Modified some imports